requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
joblib
requests
streamlit
groq
python-dotenv
jinja2
